V2.0 - C# rewrite
To make things like retrieveing a window handle (HWND) easier I rewrote everyting in C#.
As screensaver isn't a really "performance" critcal application so I think C# is better suited for that type of application.
I also intend to create a custom "editor" to create and edit the RTF files. That also seems to be way easier in C# as in C++.
With this release I also fixed a bug in the installer so that the "RichTextScreensaver.scr" file gets correctly placed into "C:\Windows\SysWOW64\TextScreensaver.scr" instead of "C:\Windows\System32\TextScreensaver.scr".